+++ src/main/java/codes/rudolph/ribac/server/error/ConflictError.java
@@ -1,8 +1,0 @@
-package codes.rudolph.ribac.server.error;
-
-public class ConflictError extends RibacError {
-
-    public ConflictError(String message) {
-        super(message);
-    }
-}