+++ src/main/java/codes/rudolph/ribac/server/error/DuplicateCreateError.java
@@ -1,0 +1,8 @@
+package codes.rudolph.ribac.server.error;
+
+public class DuplicateCreateError extends RibacError {
+
+    public DuplicateCreateError(String message) {
+        super(message);
+    }
+}