+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased] - YYYY-MM-DD
 ### Added
 * `POST /users`: Creates a new user
+* `GET /users/{userId}`: Fetches a single user
 
 ### Changed
 (changes in existing functionality)